refactor: 로그인 request에 try-catch 적용
