+++ src/js/index.js
@@ -1,5 +1,5 @@
 import '../css/index';
-import App from '../js/components/App';
+import App from './components/App';
 
 const app = new App();
